fix: unnesesary complexity fix

+++ 2_1_section/E_store.py
@@ -1,6 +1,6 @@
 def solution():
     cost, weight, money = int(input()), int(input()), int(input())
-    print(int(money - cost * weight))
+    print(money - cost * weight)
 
 
 if __name__ == "__main__":
